Use int64 when converting TIME type value
TIME type value is saved in DOUBLE, so using int64 can extend the time range.

+++ be/src/util/date_func.cpp
@@ -58,13 +58,13 @@ uint24_t timestamp_from_date(const std::string& date_str) {
     return uint24_t(value);
 }
 
-std::string time_str_from_int(int time) {
+std::string time_str_from_int(int64_t time) {
     std::stringstream time_ss;
     if (time < 0) {
         time_ss << "-";
         time = -time;
     }
-    int hour = time / 60 / 60;
+    int64_t hour = time / 60 / 60;
     int minute = time / 60 % 60;
     int second = time % 60;
